# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SML Staff Marketplace API
  version: "1.1.0"
  description: |
    Express.js port of the original Java/JAX-RS MarketPlaceWebService.

    All endpoints are under `/service/v1` (included in the server URL below).

    **Auth note**: Failed login returns HTTP 200 with `data: []` — not HTTP 401.

    **Image endpoints**: Return binary `image/png`. `/imagesguid` never returns 404 — returns empty buffer instead.
servers:
  - url: http://localhost:47309/service/v1
    description: Local development (current SML Staff Ubon)
  - url: http://localhost:47300/service/v1
    description: Local development (legacy preview)
...
security:
  - GUIDHeader: []
components:
  securitySchemes:
    GUIDHeader:
      type: apiKey
      in: header
      name: GUID
      description: Session identifier required by most endpoints
  schemas:
    ErrorResponse:
      type: object
      properties:
        ERROR:
          type: string
          example: "database connection error"
    ErrorResponseLower:
      type: object
      properties:
        error:
          type: string
    SuccessOnly:
      type: object
      properties:
        success:
          type: boolean
    EmployeeUser:
      type: object
      properties:
        user_code:
          type: string
          example: "ADMIN"
        user_name:
          type: string
          example: "ผู้ดูแลระบบ"
        permissions:
          type: array
          description: Permission keys resolved from sml_staff_permission; empty/missing table defaults to all permissions.
          items:
            type: string
          example: ["sell.view", "purchase.pu.view", "purchase.pu.create"]
    CustomerUser:
      type: object
      properties:
        user_code:
          type: string
          example: "C00001"
        user_name:
          type: string
          example: "บริษัท ABC จำกัด"
        address:
          type: string
        telephone:
          type: string
        tax_id:
          type: string
          description: From ar_customer_detail, "" if not set
...
paths:
...
  /loginemp:
    get:
      tags: [Auth]
      summary: Login พนักงาน
      description: |
        ตรวจสอบจาก erp_user (code case-insensitive, password case-sensitive)

        **หมายเหตุ**: credential ผิด → HTTP 200 + `data: []` ไม่ใช่ 401
      security: []
      parameters:
        - { name: user_code, in: query, required: true, schema: { type: string }, example: "ADMIN" }
        - { name: password,  in: query, required: true, schema: { type: string }, example: "1234" }
      responses:
        '200':
          description: Login สำเร็จ (data มี 1 element) หรือผิด (data = [])
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmployeeUser'
              examples:
                success:
                  value: { success: true, data: [{ user_code: "ADMIN", user_name: "ผู้ดูแลระบบ", permissions: ["sell.view", "purchase.pu.view"] }] }
                failed:
                  value: { success: true, data: [] }
        '400':
          description: Database error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /logincus:
    get:
      tags: [Auth]
      summary: Login ลูกค้า
      description: |
        ตรวจสอบจาก ar_customer — รหัสผ่านเก็บใน field `country` (legacy Java design)

        credential ผิด → HTTP 200 + `data: []`
      security: []
      parameters:
        - { name: user_code, in: query, required: true, schema: { type: string }, example: "C00001" }
        - { name: password,  in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: Login สำเร็จหรือผิด (ดู data[])
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomerUser'
        '400':
          description: Database error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /getEmployeePermissions:
    get:
      tags: [Auth]
      summary: สิทธิ์ของพนักงาน
      parameters:
        - { name: user_code, in: query, required: true, schema: { type: string }, example: "001" }
      responses:
        '200':
          description: Permission keys for employee
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { type: string }
        '400':
          description: user_code is required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }
        '500':
          description: DB error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }
  /setEmployeePermissions:
    post:
      tags: [Auth]
      summary: บันทึกสิทธิ์พนักงาน
      description: บันทึกลง `sml_staff_permission`; key ที่ไม่อยู่ใน registry จะถูก ignore.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_code, permissions]
              properties:
                user_code: { type: string, example: "001" }
                permissions:
                  type: array
                  items: { type: string }
                  example: ["purchase.pu.view", "purchase.pu.create"]
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessOnly' }
        '400':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }
        '500':
          description: DB error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }

  # ─── PRODUCT ─────────────────────────────────────────────────────────────────
...